salink: Move the queue management code into bank 44
To make room for more exciting code in link_obj.c.
