+++ src/queued_files.h
@@ -1,0 +1,25 @@
+/*
+ * Copyright contributors to Specasm
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
+#ifndef QUEUED_FILES_H
+#define QUEUED_FILES_H
+
+#include "salink.h"
+
+void salink_add_queued_file_e(const char *base, const char *prefix,
+			      specasm_line_t *line);
+
+#endif
