+++ src/salink.h
@@ -108,6 +108,7 @@ extern char scratch[SPECASM_MAX_SCRATCH];
 extern salink_label_t labels[MAX_LABELS];
 extern salink_global_t globals[MAX_GLOBALS];
 extern char error_buf[(SPECASM_LINE_MAX_LEN * 3) + 1];
+uint8_t salink_check_file(const char *fname);
 unsigned int salink_find_local_label_e(const char *str, int len,
 				       salink_obj_t *obj);
 unsigned int salink_find_global_label_e(const char *str, salink_obj_t *obj);
@@ -140,4 +141,7 @@ extern uint8_t obj_file_count;
 extern uint8_t obj_files_order[MAX_FILES];
 extern char image_name[MAX_FNAME + 1];
 extern unsigned int bin_size;
+extern uint8_t queued_files;
+extern const char *empty_str;
+extern const char *specasm_str;
 #endif
